refactor: extract duplicated code in cpService & mvService and create module helpers/fileOperations for them
